refacto: preprocessor + data loader DB
