+++ doctr/datasets/loaders/detection/__init__.py
@@ -1,2 +1,1 @@
 from .core import *
-from .db import *
